+++ sample/transactions/GetTransaction.php
@@ -13,7 +13,7 @@ require __DIR__ . '/../bootstrap.php';
 /// ### Retrieve Transaction
 // (See bootstrap.php for more on `ApiContext`)
 try {
-    $transaction = \BlockCypher\Api\Transaction::get('f854aebae95150b379cc1187d848d58225f3c4157fe992bcd166f58bd5063449', $apiContext);
+    $transaction = \BlockCypher\Api\Transaction::get('f854aebae95150b379cc1187d848d58225f3c4157fe992bcd166f58bd5063449', array(), $apiContext);
 } catch (Exception $ex) {
     ResultPrinter::printError("Get Transaction", "Transaction", 'f854aebae95150b379cc1187d848d58225f3c4157fe992bcd166f58bd5063449', null, $ex);
     exit(1);