+++ release_notes.md
@@ -13,6 +13,7 @@ v0.6.0
 * Added URL params 'unspentOnly' and 'before' to Address
 * Added URL params 'unspentOnly' and 'before' to FullAddress
 * Added URL params 'txstart' and 'limit' to Block
+* Added URL params 'instart', 'outstart' and 'limit' to Transaction
 
 v0.5.0
 ------